Test --id-map where the code it exercises already is
The manifest tests read the DB back and compare against store.id_columns and
synth, so they belong with the loader's own tests rather than in a file of their
own. The CLI half is separate: --id-map is in _BYO_ONLY, so test_cli.py is what
should know it arrives as a Path, appears in --help, and is refused under the
bench type.

The bench-type refusal was parametrized over bare flags, which only works for
boolean ones — an option that takes a value stops at the missing argument
instead. Parametrizing over argv fragments covers --roster too, which was
missing for the same reason.
